Forgot to also rename web_safe_test back to just vm_test
